modify balance table amount from int to bigint

+++ database/dump.sql
@@ -49,7 +49,7 @@ CREATE TABLE `balances` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `address` varchar(256) NOT NULL,
   `asset_id` char(64) NOT NULL,
-  `amount` int(19) DEFAULT '0',
+  `amount` bigint(20) DEFAULT '0',
   `created_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (`id`),
   KEY `address` (`address`),
